Bites

Documentation for Bites.

Bites.bite_steps

Bites.calculate_r0

Bites.distribute_bite_probabilities

Bites.bite_steps — Method

bite_steps(n_steps::Int64, n_humans::Int64, n_mosquitoes::Int64, human_infection_time::Int64, mosquito_life_span::Int64, human_probs::Array{Float64, 1}, mosquito_probs::Array{Float64, 1}, transmission_prob::Float64)::Tuple{Array{Int64, 1}, Array{Int64, 1}, Array{Int64, 1}}

Returns a Tuple containing (1) an array with number of infected mosquitoes at each time step, (2) an array with number of infected humans at each time step, and (3) an array with number of recovered humans at each time step.

Bites.calculate_r0 — Method

  calculate_r0(n_reps::Int64, infection_ts::Array{Int64,2}, seed_cases::Int64)

Returns a named Tuple containing (1) R0 (mean of reps), (2) R0 for each repetition, (3) mean R0 for successive numbers of repetitions.

Parameters

n_reps The number of repetitions in the simulation. In64.

infection_ts The matrix time series of number of infected cases at each time step. A two-dimensional array of In64 with dimensions equal to number of repetitions and number of time steps.

seed_cases Number of infected cases seeded into the population at the beginning of each repetition. If seed_cases=0, then the first cases will be detected in the data. (For example, if we seed 1 infected human case into the population, then we would use seed_cases=1 for the huiman R0 and seed_cases=0 for the mosquito R0.)

Bites.distribute_bite_probabilities — Method

distribute_bite_probabilities(D1, D2, N1::Int64, N2::Int64, expected_bites::Float64)::Tuple{Array{Float64, 1}, Array{Float64, 1}}

Returns a Tuple containing (1) per mosquito probabilities of being bitten for each person, and (2) per person probabilities of biting for each mosquito, with the per mosquito probabilities of being bitten for each person adjusted to the desired expected bites value.

Parameters

D1 The distribution from which the human probabilities of being bitten should be drawn. A distribution from the Distributions.jl package.

D2 The distribution from which the mosquito probabilities of biting should be drawn. A distribution from the Distributions.jl package.

N1 The number of humans in the population. Int64.

N2 The number of mosquitoes in the population. Int64.

expected_bites The expected number of total bites (total links in the bipartite network). Float64.